requirements.txt
streamlit
langchain
langgraph
groq
langchain-groq
psycopg2-binary
pgvector
sqlalchemy
python-dotenv
pillow